+++ base/src/MVL1.h
@@ -6,25 +6,25 @@
 
 namespace Falm {
 
-double L0Dev_DotProduct(Matrix<double> &a, Matrix<double> &b, Mapper &pdom, Mapper &map, dim3 block_dim);
+REAL L0Dev_DotProduct(Matrix<REAL> &a, Matrix<REAL> &b, Mapper &pdom, Mapper &map, dim3 block_dim);
 
-double L0Dev_Norm2Sq(Matrix<double> &a, Mapper &pdom, Mapper &map, dim3 block_dim);
+REAL L0Dev_Norm2Sq(Matrix<REAL> &a, Mapper &pdom, Mapper &map, dim3 block_dim);
 
-double L0Dev_MaxDiag(Matrix<double> &a, Mapper &pdom, Mapper &map, dim3 block_dim);
+REAL L0Dev_MaxDiag(Matrix<REAL> &a, Mapper &pdom, Mapper &map, dim3 block_dim);
 
-void L1Dev_ScaleMatrix(Matrix<double> &a, double scale, dim3 block_dim);
+void L1Dev_ScaleMatrix(Matrix<REAL> &a, REAL scale, dim3 block_dim);
 
-static inline double L1Dev_DotProduct(Matrix<double> &a, Matrix<double> &b, Mapper &pdom, dim3 block_dim) {
+static inline REAL L1Dev_DotProduct(Matrix<REAL> &a, Matrix<REAL> &b, Mapper &pdom, dim3 block_dim) {
     Mapper map(pdom, Gd);
     return L0Dev_DotProduct(a, b, pdom, map, block_dim);
 }
 
-static inline double L1Dev_Norm2Sq(Matrix<double> &a, Mapper &pdom, dim3 block_dim) {
+static inline REAL L1Dev_Norm2Sq(Matrix<REAL> &a, Mapper &pdom, dim3 block_dim) {
     Mapper map(pdom, Gd);
     return L0Dev_Norm2Sq(a, pdom, map, block_dim);
 }
 
-static inline double L1Dev_MaxDiag(Matrix<double> &a, Mapper &pdom, dim3 block_dim) {
+static inline REAL L1Dev_MaxDiag(Matrix<REAL> &a, Mapper &pdom, dim3 block_dim) {
     Mapper map(pdom, Gd);
     return L0Dev_MaxDiag(a, pdom, map, block_dim);
 }
